refactor(saga): extract _StepExecutor into separate module
Move the internal _StepExecutor helper class from sagaz/core/saga.py
into a dedicated module (sagaz/core/_step_executor.py) to improve file
organization and reduce saga.py complexity.

- Reduces saga.py: 1285 → 1251 lines (-34 LOC)
- Preserves all exports; _StepExecutor remains internal (prefixed `_`)
- Adds comprehensive docstrings in new module
- All 506 core unit tests pass

This addresses Tier 2 size outlier reduction for core/saga.py.

Relates-to: #168
